+++ es6/tests/fixtures.js
@@ -446,7 +446,7 @@ const fixtures = [
 				file: "word/document.xml",
 				id: "scopeparser_execution_failed",
 				scope: 1,
-				tag: "b",
+				xtag: "b",
 				offset: 4,
 			},
 		}),
@@ -2034,6 +2034,24 @@ fixtures.push({
 	it: "should work with rawxml with undefined tags",
 	scope: {},
 	result: "BEFOREAFTER",
+});
+fixtures.push({
+	...rawXmlTest,
+	it: "should throw error with rawxml with scope that is an integer",
+	scope: { rawxml: 11},
+	errorType: Errors.XTTemplateError,
+	error: wrapMultiError({
+		name: "RenderingError",
+		message: "Non string values are not allowed for rawXML tags",
+		properties: {
+			explanation: "The value of the raw tag : 'rawxml' is not a string",
+				file: "word/document.xml",
+				id: "invalid_raw_xml_value",
+				value: 11,
+				xtag: "rawxml",
+				offset: 0,
+		},
+	}),
 });
 
 fixtures.forEach(function (fixture) {